test: 각 UseCaes별로 파일 분리
- 각 파일별 SystemUnderTest를 정의하여 테스트에 용이하게 변경
