+++ ChildOfWeatherTests/DetailShowUseCaseTests.swift
@@ -1,0 +1,23 @@
+//
+//  DetailShowUseCaseTests.swift
+//  ChildOfWeatherTests
+//
+//  Created by 조영민 on 2022/04/18.
+//
+
+import XCTest
+@testable import ChildOfWeather
+
+class DetailShowUseCaseTests: XCTestCase {
+
+    override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+
+
+}
